+++ src/fable-standalone/test/bench-compiler/app.fs
@@ -260,7 +260,6 @@ let run opts projectFileName outDir =
         sourceMaps = (opts |> hasFlag "--sourceMaps") || (opts |> hasFlag "-s")
         typedArrays = opts |> tryFlag "--typedArrays"
         language = opts |> argValue ["--language"; "--lang"]
-                        |> Option.orElse (tryFlag "--typescript" opts
                         |> Option.map (fun _ -> "TypeScript"))
                         |> Option.defaultValue "JavaScript"
         printAst = opts |> hasFlag "--printAst"
@@ -281,7 +280,6 @@ Options:
   --typedArrays     Compile numeric arrays as JS typed arrays (default is true for JS, false for TS)
   --run             Execute the script after compilation
 
-  --typescript      Compile to TypeScript (experimental)
   --optimize        Compile with optimized F# AST (experimental)
 """